[FEATURE] 기관 조회수 기능 구현 #188
- 기관 상세 조회 시 조회수 증가

+++ module-user/src/main/java/org/example/domain/institution/controller/response/ListInstitutionDto.java
@@ -20,5 +20,5 @@ public class ListInstitutionDto {
   private String name;
 
   @Schema(description = "조회수")
-  private long viewCount;
+  private int viewCount;
 }
